Adding Stub, renaming classes
